binary_segmentation use cuda
adds a cuda implmentation of the binary segmentation,
factors dispatch of cpu or gpu into an internal namespace.
