+++ alg/CMakeLists.txt
@@ -11,6 +11,7 @@ set(teca_alg_cxx_srcs
     teca_apply_binary_mask.cxx
     teca_apply_tempest_remap.cxx
     teca_binary_segmentation.cxx
+    teca_binary_segmentation_internals.cxx
     teca_bayesian_ar_detect.cxx
     teca_bayesian_ar_detect_parameters.cxx
     teca_cartesian_mesh_coordinate_transform.cxx
@@ -58,6 +59,13 @@ set(teca_alg_cxx_srcs
     teca_dataset_diff.cxx
     )
 
+set(teca_alg_cuda_srcs)
+if (TECA_HAS_CUDA)
+    set(teca_alg_cuda_srcs
+        teca_binary_segmentation_internals.cu
+        )
+endif()
+
 set(teca_alg_f90_srcs)
 
 set(teca_alg_f90_generics
@@ -91,12 +99,12 @@ if (TECA_HAS_BOOST)
     list(APPEND teca_alg_link ${Boost_LIBRARIES})
 endif()
 
-set(teca_alg_cuda_srcs)
 if (TECA_HAS_CUDA)
     set_source_files_properties(${teca_alg_cxx_srcs} PROPERTIES LANGUAGE CUDA)
 endif()
 
-add_library(teca_alg ${teca_alg_cxx_srcs} ${teca_alg_f90_srcs})
+add_library(teca_alg ${teca_alg_cxx_srcs} ${teca_alg_cuda_srcs} ${teca_alg_f90_srcs})
+
 target_link_libraries(teca_alg teca_data teca_core ${teca_alg_link})
 
 target_include_directories(teca_alg
